Re-namespace Counter
