+++ osu.Framework.Tests/Visual/TestCaseCountingText.cs
@@ -80,38 +80,4 @@ namespace osu.Framework.Tests.Visual
 
         protected override void OnCountChanged() => text.Text = resultFunction(Count);
     }
-
-    public class Counter : CompositeDrawable
-    {
-        private double count;
-        /// <summary>
-        /// The current count.
-        /// </summary>
-        protected double Count
-        {
-            get { return count; }
-            private set
-            {
-                if (count == value)
-                    return;
-                count = value;
-
-                OnCountChanged();
-            }
-        }
-
-        /// <summary>
-        /// Invoked when <see cref="Count"/> has changed.
-        /// </summary>
-        protected virtual void OnCountChanged() { }
-
-        public TransformSequence<Counter> CountTo(double endCount, double duration = 0, Easing easing = Easing.None)
-            => this.TransformTo(nameof(Count), endCount, duration, easing);
-    }
-
-    public static class CounterTransformSequenceExtensions
-    {
-        public static TransformSequence<Counter> CountTo(this TransformSequence<Counter> t, double endCount, double duration = 0, Easing easing = Easing.None)
-            => t.Append(o => o.CountTo(endCount, duration, easing));
-    }
 }
